requirements/test.txt
atomicwrites==1.3.0
attrs==19.1.0
bandit==1.5.1
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
defusedxml==0.5.0
django-allauth==0.39.1
django-filter==2.1.0
django-guardian==1.5.0
django-rest-auth==0.9.5
drf-writable-nested==0.5.1
entrypoints==0.3
filelock==3.0.10
flake8==3.7.7
funcy==1.11
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
mccabe==0.6.1
more-itertools==7.0.0
oauthlib==3.0.1
pbr==5.1.3
pipdeptree==0.13.2
pluggy==0.9.0
psycopg2==2.7.7
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytest==4.4.0
pytest-django==3.4.8
python-dotenv==0.10.1
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
smmap2==2.0.5
sqlparse==0.3.0
stevedore==1.30.1
toml==0.10.0
tox==3.8.6
urllib3==1.24.1
virtualenv==16.4.3
